Re organized some struct for easy access
