+++ src/content/game_content/interface/storage.rs
@@ -1,8 +1,9 @@
 use graphics::*;
 
 use crate::{
-    data_types::*, is_within_area, logic::*, socket::sends::*, widget::*,
-    Alert, AlertIndex, AlertType, Interface, Result, Socket, SystemHolder,
+    Alert, AlertIndex, AlertType, Interface, Item, Result, Socket,
+    SystemHolder, data_types::*, is_within_area, logic::*, socket::sends::*,
+    widget::*,
 };
 
 use super::ItemDescription;
