+++ src/systems/renderer/fade.rs
@@ -3,7 +3,8 @@ use graphics::*;
 use winit::dpi::PhysicalSize;
 
 use crate::{
-    BufferTask, Result, Socket, SystemHolder, World, content::*, data_types::*,
+    BufferTask, Entity, Position, Result, Socket, SystemHolder, World,
+    content::*, data_types::*,
 };
 
 #[derive(Default)]
@@ -251,8 +252,9 @@ pub fn fade_end(
             content.switch_content(world, systems, ContentType::Game)?;
 
             let pos = if let Some(entity) = content.game_content.myentity {
-                if content.game_content.in_game {
-                    world.get_or_err::<Position>(entity)?
+                if let Some(Entity::Player(p_data)) = world.entities.get(entity)
+                {
+                    p_data.pos
                 } else {
                     Position::default()
                 }
